changes in search and edit books
Big changes, you can now search and edit books. also GUIComponents.java is where duplicate components are used for other pages. Different for CustomLayout
