+++ LandingPagesGUI/CustomLayoutManager.java
@@ -128,9 +128,9 @@ public class CustomLayoutManager {
         return searchBar;
     }
 
-    // Create an input field,
+    // Create an input field, parameter panel to where to store the input field
     // has parameter JLabel to customize what type of input field(ex: username)
-    public void createInputField(JPanel panel, JLabel label) {
+    public JTextField createInputField(JPanel panel, JLabel label) {
         System.out.println(this + " > Created an Input Field for " + label.getText());
         panel.add(label);
 
@@ -139,6 +139,8 @@ public class CustomLayoutManager {
 
         labelStyleDefault(label); // change the style of label
         textfieldStyleDefault(textField); // changes text field style(lol)
+
+        return textField;
     }
 
     // LABEL STYLES
